[ADD] test_utm: add a testing module for UTMs
This commit introduces a "test_utm" module to be able to test a complete UTM
flow without depending on existing functional models.

Defining testing models is only possible within a new module.

The goal is to make sure the UTM values assigned through cookies are properly
set on the resulting created record.
Along with auto-creation and clean fallback on invalid values.

Task-3330036