[REF] mail: _mail_get_operation_for_mail_message_operation using Domain
Instead of filtering the recordset, the function returns now a list of
pairs `(Domain, opeartion)`. We can use the domain to filter records
to which the operation applies. This way we do not have to fetch the
records before filtering them and will be able to use it in SQL.