[PERF] mrp: speedup _compute_quantities
A bottleneck appears in mrp when _compute_quantities is called on a
large recordset. The calls to `mrp.bom.explode` can stack up
and lead to slow execution. A solution for this is to memoize the
exploded values of the boms. We can do that because this value is
actually independent of the current product, as we are not using the
first list returned by the `explode` method. This leads to a noticeable
speedup, especially in database where there are a small number of boms
for a large number of product.products.

One thing to be wary of is that some components can be applied to
differnt variants in different quantities. So there are some exceptions
to the memoization and a help function is introduced to check whether to
use the cached exploded values or to explode the current bom.

speedup

Time to search for products with `free_qty` > 0.0. Database with mrp
installed and kit boms. Most of the kit boms are for templates. 800 000
products in the db, 15 000 templates.

| Total number of products | Before PR | After PR  |
|:------------------------:|:---------:|:---------:|
|           100            |    51ms   |   48ms    |
|          1000            |    220ms  |   200ms   |
|         10000            |    1.2s   |   900ms   |
|        100000            |    10s    |   5.77s   |
|        800000            |    15min  |   1min15s |

The customer database has 3 boms for templates that have more than
20 000 variants each. In this case, memoizing the explode gives a
significant speedup. That's what happens in the 800 000 case above.

opw-4948266