[IMP] mrp: replace days to prepare computation with popover
Currently, users must rely on a manual computation to determine the Days to
prepare for a BOM. To access this information, users need to click the 'Compute'
link, which adds unnecessary friction to the workflow.

Moreover, the existing computation only provides the number of days required for
preparation, without explaining which component or BOM contributes to the lead
time. In many cases, identifying the root cause of the delay becomes difficult
during MRP operations.

This PR replaces the Days to Prepare computation with a contextual popover,
offering a more user-friendly and informative experience. Information in the
popover is computed automatically when opening a BOM, eliminating the need for
manual interaction.

- The popover highlights the component contributing the longest lead time. If
  the longest lead time comes from a child Manufacturing Order (MO), the popover
  displays the name of that MO; otherwise, it shows the component name
  responsible for the delay. It also displays the route information associated
  with the component.
- The popover is computed only for BOMs of type Normal and Subcontracting. It is
  not computed and remains invisible for Kit BOMs.
- If no manufacturing route is defined for a component, the popover view will
  be shown as 'Not Available: Missing Route Info for Components.' and Clicking
  'Check BoM overview' redirects the user to the BoM overview of the
  corresponding product in forecast mode.
- Removed the redundant action button 'Days to prepare Manufacturing Order'
  from the Miscellaneous Page, since the new fields cover that functionality.
  Updated related test cases to align with the new logic.
- Renamed 'Days to Prepare Manufacturing Order' to 'Days to Prepare' for better
  clarity and to avoid confusion.

This improvement streamlines BOM lead-time visibility by replacing the previous
approach with a more intuitive and insightful one. It also provides users with
clear information about which component contributes to the lead time, improving
production planning efficiency and reducing unnecessary clicks.

Task ID:5022350