[IMP] *: modify weight and volume in packages
*: stock, stock_picking_batch, stock_delivery

This commit improves the computation of weight and volume in packages.

This commit:
1- refactors the code by moving the logic of a package's weight
and volume computation into `stock` module instead of `stock_delivery`.

2- fixes the computation of the picking weight to include
the base_weight of the package_type for all layers of the package
in the picking and in the batch transfers.

3- adds a computed `volume` field on the package. It is calculated using
the package_type dimensions if exist, otherwise it uses the quant_ids.

4- modifies the picking volume computation to take into account the
`shipping_volume` of the package if exists (instead of the products),
otherwise it computes the volume of what's inside the package, whether
it includes other packages or products.

Related upgrade PR: https://github.com/odoo/upgrade/pull/8749

Task-5056474