[IMP] html_builder, *: allow translation of image src
*: html_editor, website, tools

This commit allows to translate an image's src, making it possible to
display a different image in language A and language B.

In order to avoid having src URLs wrapped inside a translating <span>
(see `translate_func()`), which sends wrong requests when the DOM is
loaded, we watch the actual translatable URL string on a
`data-oe-translatable-link` attribute.

We also prevent src attributes from being exported in .pot files.

task-3626918

Co-authored-by: Guillaume Dieleman (gdi) <gdi@odoo.com>
Co-authored-by: Robin Lejeune (role) <role@odoo.com>