[FIX] hr,hr_gamification,hr_presence: employee onboarding helper
Issue:
------

The "employee no content onboarding helper" (screen with `Load sample data` button) shouldn't be shown on other companies `Employee` app than the `base.main_company`. Indeed, this button only loads the sample data for `base.main_company`.

In the process of solving this bug, this task ended up being a re-design of the no content helper.

Reproducing steps:
------------------

- Create a DB with `hr` module and without demo data
- Create a second company
- Go in the Employees app, the onboarding helper should be displayed
- Click on the "Load sample data" button, the page is reloaded, but nothing happens

Purpose:
--------

Improving the no content employee helper so that :
A) It displays "Load sample data" and "Create an employee" buttons with some onboarding design things when the sample data have not been loaded, the user has the rights to create employees, we are on `base.main_company` and the `base.main_company` has no employees.
B) Otherwise, when the company has no employee, it displays an empty screen
Moreover, the empty screen should contain a "New Employee" button if the user has the rights to create employees.

Same for mobile, but the (A) view comes without the onboarding design things if there is not enough space.

community-229238
task-4879557