[IMP] survey_crm: cleanup survey_crm "origin" fields
RATIONALE

The way we currently use the UTM trackers (aka source/medium/campaign) does not
comply with the industry standards, leading to confusion and globally unclear
or not usable reporting on UTM data.

PURPOSE

Now that we have a global concept of UTM reference, we can use it to replace
the recently introduced field: origin_survey_id.

By adapting the various domain and actions, we end up getting the same
functional result without requiring this field.

(Note: "origin_survey_ids" was unused and cleaned at the same time).

Task-3330036