[REM] website: remove interaction that added d-none on s_popup
The interaction `shared_popup` synchronised the class `d-none` on the
popup element with the visibility of the modal inside.

Since the commit 15c6abe6c518389e367cde2aa944bbb511519571, which adds
a css rule to force the height of the `s_popup` of the cookie bar to 0,
this is not needed anymore.

This commit removes the part of the `shared_popup` interaction about the
synchronization. It instead just removes the `d-none` on the popups that
were saved with it.

task-5149984