[IMP] website, *: dynamic snippet products with title on left
*=website_sale

This commit ensures that two items are displayed side-by-side in
"s_dynamic_snippet_products" when the section title is placed on the
left and the content-width is not max (container-fluid). A new
"builderAction" updates the numberOfElements accordingly, and the
content-width option plugin is patched to keep this value consistent
when the width changes. Additionally, some unnecessary product snippet
related CSS has been removed from "000.scss" file of "s_dynamic_snippet"
to achieve the correct side-by-side layout in this scenario.

task-5090468