[IMP] [Attendance] Load the Overview Page Faster
The Gantt page was taking a long time to load when having a 1000+ employees
I enabled pagination for the gantt view to make it load less fields (and
not everything at once).
Previously, active employees were loaded separately, with a second
get_gantt_data() call.
Merging the two calls into one was already an
improvement. Also, the previous way of calling get_gantt_data() twice
did not respect the limit and offset logic.

In community, _read_group_employee_id would return all employees when
"gantt_start_date" was in the context. This was the line slowing the
request the most. Moreover, the result of _read_group_employee_id would
just be discarded if the length of the result was greater than the
limit. So it would be just slowing the request for nothing.
Removing entirely that logic and moving it up in the call chain
(in get_gantt_data) allowed me to add those employees without any
record, while also respecting the limit and offset parameters.

task-4762235