[FIX] website: redirect to the right URL after saving SEO modifications
Previously, the value of the Custom Url field replaced the object's seo_name
in the url of the redirection. However, if this value was null, the resulting
url might not match any of the controllers' URL patterns. This change allows
this replacement to occur only when a value has been entered in the Custom Url
field.

Task-5114394