[FIX] point_of_sale: load all pos categories
Since this PR https://github.com/odoo/odoo/pull/225658, if the POS category of a combo item product was not listed under the restricted categories but was assigned to a preparation printer, the item would not be printed by the preparation printer.

A test was already present to check this, but the validation method was incorrect  (test_restricted_categories_combo_product)

This commit ensures that all POS categories are loaded, preventing any category from being missed.

X-original-commit: b9cf922f9f4a29bc2e6d7d51c9698e8ff97a85c6