[IMP] account: hash cash journal on reco
The mechanism for securing accounting entries already exists in most of our
journals. However, it has not yet been implemented for ‘Cash’ journals.

This commit will allow users to check the "Secure Posted Entries with Hash"
field. Now when this is checked, and the entry is reconciled. The entry will be
hashed.

For cash journal only, we can have draft statement.

Also adding a cash suspense account, on posting if a statement line (in a
statement) has no account, we will set this one.

task-5170039