[IMP] mass_mailing: add dynamic snippet plugin
This commit adds a plugin that will manage snippets with dynamic
record content pulled straight from Odoo.

The plugin will fetch record data (such as names and dates) and add
it as a fragment to the snippet. Users will then be able to edit the
fragment at their convenience.

Applications for the website_event and sales modules will follow.

The mass_mailing background option plugin is modified to work on
these dynamic snippets' images.

task-4247666