[FIX] pos_event: fix seats availabilities computation
Purpose
=======
Fix the event, slot and ticket availabilities computation to make sure
the correct availabilities are taken into account when registering to an event.

Specification
=============
Previously the event registrations already validated during the current order are
taken into account but only in the ticket seats available display.
This is an issue as the slot maximum number of seats is not considered meaning
the user is misleaded and could book a ticket even though it'll exceed the
limitations making the order crash at payment step.

=> Fixing the issue by dynamically updating the availabilities per slot-ticket, per slot and
per ticket depending on the already validated registrations for the current order.

Previously the availability per slot was computed using a sum on the slot-ticket availabilities.
This is wrong as the slot-ticket availabilities are accurate only if the user registers
to them one slot-ticket at a time. Here the UI allows multiple ticket selection, so the sum of the
slot-ticket availabilities doesn't represent the actual slot availabilities as it could exceed
the event/slot limitations.

=> Fixing the issue by correctly taking into account the slot and ticket limitations when
computing the slot availabilities.

Task-5079357