[FIX] pos_event: fix small slot and ticket limitations issues
Purpose
=======
Fix some small issues regarding the slot and ticket seats limitations when
booking an event from the pos.

Specification
=============
1) Fix the ticket seats available display for non-multi slot events.

When the event seats are limited (seats_max > 0 and seats_limited=True)
but the ticket seats aren't limited (ticket.seats_max == 0),
the ticket available seats shouldn't be based on the ticket.seats_available field
(which is only relevant when the ticket seats are limited) but should be
based on the event available seats instead.

This issues doesn't happen for multi slots events as in this case the
'get_slot_tickets_availability_pos' python method is called returning the correct
ticket availabilities.

2) Make sure to check if the user selected more tickets than the available event/slot seats.
It's not checked and crashes at payment step.

When confirming the event configurator popup, each ticket quantity is checked based
on their own limitations (ticket available seats). However the total number of
selected tickets could be limited by the event or the slot as well (event/slot available seats).
Adding a check to make sure the total number of tickets doesn't exceed the event/slot limitations.

3) Make sure to display the correct number of seats available/the correct slot count on the product card.

Currently, the number of available seats for non-multi slots events on the product card isn't correctly computed.
When the event has limited seats and at least one unlimited ticket, the computation should use the event
available seats. Instead of that, it uses the sum of every ticket seats available even though this field is only
relevant when the tickets are limited.

Currently the slot count on the product card is the one of the future slots but some of those slots
could be unavailable (their button disabled on the slot selection popup) and shouldn't be included
in the product slot count.

4) Do not raise error when trying to register to a multi-slots event without any ticket max seats
limitations.

Currently, if not a single ticket has an availability > 0, a danger notification is raised
preventing the registration to the event. However, the availability is not necessarily an integer,
it could be equal to "unlimited" and in this case the danger notification will appear even though
there are available tickets. Fixing the issue by evaluating "unlimited" tickets to be available.

5) Dynamically adapt the number of tickets left depending on the ticket selection.

Currently, on the event configuration page, there's a counter with the number of tickets left
under each ticket name. This number is never updated no matter the selected tickets.
The only indicator to know if there's at least one ticket left is a color indicator.
The color indicator is not enough, if there is 5 tickets left, selecting 3 tickets will need
to update the counter accordingly to 2 so that the user knows what is happening.

Task-5079357