[FIX] website, web_editor: only set lang in website xml editor
Problem: When opening the Studio XML editor when Website is installed, the
translation terms corresponding to the Default Language of the first
website in the database are used. This behavior should only be applied
to the HTML/CSS Editor in Website.

Purpose: Modify Website's override of get_related_views to only apply
lang when it is called from get_assets_editor_resources

Steps to Reproduce in Runbot:

1. Activate a non-English (US) language.
2. Add this language to the Website with the lowest ID in the database,
   then set it to the Default Language of the Website.
3. Enter Studio and navigate to a view that has translation terms in its
   view (ex. Invoice PDF Report), then open the XML editor.

opw-5136124