[IMP] mail: always search display_name and email in recipients input
The recipients input only searches emails if the input is identified as
an email (e.g. must contain an @). Instead always search based on email.

Additionally searching display_name can be useful to search based on
company. And searching on email instead of email_normalized allows
finding recipients who don't have valid emails, which can be fixed
afterwards.

The company is now shown as part of the search results so the results
are consistent, using field-specific odoomark to avoid clutter.

Emails are now also shown in the full composer to be consistent.

task-5013894