[FIX] *: better translations in templates
This commit aims to provide better translatable strings from static XML
templates.

To do so, it includes the following changes:
- isolated strings that are not meaningful to a translation, such as
punctuation or special characters (e.g. "?" for field tooltip marker,
"-" to separate tips, or a "x" button to close a panel);
- add `t-translation="off"` to strings that are not supposed to end up
in translatable strings, such as the aforementioned characters;
- a few simplifications for the affected templates have also been
applied to the changed nodes when possible, to enforce semantics or to
improve readability.