[IMP] mrp: component consumption warning wizard improvements
The consumption warning wizard was revamped. The functional changes are
as follows:

1. The wizard previously only kicked in if the MO had a BoM. Now it
works without a BoM as well, comparing the consumed quantity against the
expected quantity of the move (specified by the user) instead.

2. The wizard no longer lumps together distinct components of the
same product (say, with different UoMs) and treats each move separately.

3. If a BoM line is missing the corresponding component move, it tries
to match it with a component that likewise lacks a corresponding BoM
line, in case the user removed and readded the move.

4. The wizard now displays and handles the moves by the UoMs specified
in the MO rather than the UoMs set for the moves' respective products.

5. Kits are handled as one would expect: It acts like the MO has a
regular BoM comprising the BoM lines resulting from flattening the
nested kits of the components of the MO's BoM (ie what the MO does when
it automatically explodes the kits when the user confirms the order) and
assumes the resulting moves are user-specified values in absence of the
MO's BoM.

Shortcomings:

1. Right now the heuristics to pair orphan BoM lines with orphan
component moves only checks by the product id and the unit of measure
(if enabled). It could be replaced with something more clever.

2. Component quantities with fine fractional values can lead to false
positives (and, extremely rarely, false negatives). This is existing
behaviour and not a regression.

Task ID: 4603609, 4441090