[IMP] mail: move edit near delete for other user messages
For messages of other users, even if I am technically allowed to edit
the message it is much less common than simply wanting to reply.

Instead move the "edit" action near the "delete" action when a message
was not written by the current user, such that "reply all" becomes the
last "quick action" on desktop.

task-5013894