[FIX] website, tools: make `<select>` options individually translatable
__Current behavior before commit:__
[This PR] made the `<select>` field translatable by adding `select`
and `option` to `TRANSLATED_ELEMENTS` and by implementing a temporary
element to handle the frontend translation.

Consequently, the method `translate_xml_node` considers the `<select>`
field translatable as a whole. Editing the text of one option therefore
resets the translation of all the other options as well.

__Description of the fix:__
- Remove `select` and `option` from `TRANSLATED_ELEMENTS` so each
option is considered as an independent translatable term.
- Adapt the temporary element hack accordingly; the translation wrapper
`<span>` is put inside an attribute of the options because HTML only
allows plain text inside `<option>` elements.
- Remove the now useless `SelectTranslateDialog`.

__Steps to reproduce:__
1. Add a form snippet to a website page.
2. Add a selection field to the form.
3. Translate the options in another language.
4. Leave Translation mode and go back to Edit mode.
5. Edit one of the options.
6. Change the website back to the second language.
7. Open Translation mode to see the delayed translations.
=> All options in the select field are reset.

[This PR]: https://github.com/odoo/odoo/pull/117519

task-5116823