[IMP] mail: use reply_to for forward and replies
The "reply all" and "forward" buttons on messages open a composer with
the message quoted and attributed to the original author.

In most cases it is more relevant to set the "reply-to" email there as
previous recipients will have interacted with that address. The author
may not want their address "leaked". "email_from" is reordered to take
priority for the same reason.

task-5013894