[IMP] mail: reuse the subject
People often want to set some recipient-friendly name for the subject of
a record. However the only way to set a "default subject" is usually to
modify the name of the record, which is not always reasonable.

We add a concept of "default subject" field which gets updated whenever
a different subject is used. This allows users to set a subject once and
never have to set it again.

Additionally it is now possible to set the subject from the chatter
directly, and see what the default is from there.

task-5013894